requirements.txt
aiogram[redis]==3.15.0
aiosqlite==0.20.0
apscheduler==3.10.4
python-dotenv==1.0.1
aiohttp==3.11.11
yookassa==3.3.0
redis==5.2.1
